+++ core/stencil.config.ts
@@ -155,6 +155,7 @@ export const config: Config = {
         'ion-loading',
         'ion-modal',
         'ion-picker',
+        'ion-popover',
         'ion-toast',
         'ion-toast',
